requirements.txt
https://github.com/nccr-itmo/FEDOT/archive/composer-history-fix2.zip
flask-socketio==5.0.1
Flask==1.1.1
flask-accepts==0.10.0
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-Script==2.0.6
marshmallow==3.2.0
Werkzeug==0.16.1
flask-restx==0.2.0
tensorflow>=2.2.0
pytest>=5.2.0
pytest-flask==1.1.0
Flask-Login==0.4.1
Flask-SQLAlchemy==2.5
oauthlib==3.0.1
pyOpenSSL==19.0.0
PyJWT==2.0.1
pydot==1.4.2
networkx==2.4
flask-cors==3.0.10
flask_pymongo==2.3.0
pytest_mongodb==2.2.0
dnspython==2.1.0